--- src/components/Button/index.ts
@@ -1,3 +1,0 @@
-export { ButtonCustom } from "./ButtonCustom"
-export { ButtonSwitch } from "./ButtonSwitch"
-export { ButtonIcon } from "./ButtonIcon"
