--- src/features/deposit/index.ts
@@ -1,1 +1,0 @@
-export { DepositWidget } from "./components/DepositWidget"
