--- src/features/index.ts
@@ -1,3 +1,0 @@
-export * from "./swap"
-export * from "./deposit"
-export * from "./withdraw"
