+++ src/features/machines/backgroundQuoterMachine.ts
@@ -2,7 +2,7 @@ import { type ActorRef, type Snapshot, fromCallback } from "xstate"
 import { settings } from "../../config/settings"
 import { type AggregatedQuote, queryQuote } from "../../services/quoteService"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../types/base"
-import { isBaseToken } from "../../utils"
+import { isBaseToken } from "../../utils/token"
 
 export type QuoteInput =
   | {
