+++ src/features/swap/components/SwapForm.tsx
@@ -10,7 +10,8 @@ import {
 } from "react"
 import { useFormContext } from "react-hook-form"
 import type { ActorRefFrom, SnapshotFrom } from "xstate"
-import { ButtonCustom, ButtonSwitch } from "../../../components/Button"
+import { ButtonCustom } from "../../../components/Button/ButtonCustom"
+import { ButtonSwitch } from "../../../components/Button/ButtonSwitch"
 import { Form } from "../../../components/Form"
 import { FieldComboInput } from "../../../components/Form/FieldComboInput"
 import { SwapIntentCard } from "../../../components/IntentCard/SwapIntentCard"
@@ -18,8 +19,7 @@ import type { ModalSelectAssetsPayload } from "../../../components/Modal/ModalSe
 import { useModalStore } from "../../../providers/ModalStoreProvider"
 import { isAggregatedQuoteEmpty } from "../../../services/quoteService"
 import { ModalType } from "../../../stores/modalStore"
-import type { SwappableToken } from "../../../types"
-import { isBaseToken } from "../../../utils"
+import type { SwappableToken } from "../../../types/swap"
 import { computeTotalBalance } from "../../../utils/tokenUtils"
 import type { depositedBalanceMachine } from "../../machines/depositedBalanceMachine"
 import type { intentStatusMachine } from "../../machines/intentStatusMachine"
