+++ src/features/swap/components/SwapSubmitter.tsx
@@ -1,6 +1,6 @@
 import { providers } from "near-api-js"
 import { type ReactNode, createContext } from "react"
-import type { ChainType } from "../../../types"
+import type { ChainType } from "../../../types/deposit"
 import type { SendNearTransaction } from "../../machines/publicKeyVerifierMachine"
 import { SwapUIMachineContext } from "./SwapUIMachineProvider"
