+++ src/features/swap/components/SwapUIMachineFormSyncProvider.tsx
@@ -1,7 +1,8 @@
 import { useSelector } from "@xstate/react"
 import { type PropsWithChildren, useEffect, useRef } from "react"
 import { useFormContext } from "react-hook-form"
-import type { ChainType, SwapWidgetProps } from "../../../types"
+import type { ChainType } from "../../../types/deposit"
+import type { SwapWidgetProps } from "../../../types/swap"
 import { usePublicKeyModalOpener } from "../hooks/usePublicKeyModalOpener"
 import type { SwapFormValues } from "./SwapForm"
 import { SwapUIMachineContext } from "./SwapUIMachineProvider"
