--- src/features/swap/index.tsx
@@ -1,2 +1,0 @@
-export { SwapWidget } from "./components/SwapWidget"
-export { SwapForm } from "./components/SwapForm"
