+++ src/features/withdraw/components/WithdrawForm/index.tsx
@@ -19,26 +19,26 @@ import { Fragment, type ReactNode, useEffect, useMemo } from "react"
 import { Controller, useForm } from "react-hook-form"
 import { useTokensUsdPrices } from "src/hooks/useTokensUsdPrices"
 import type { ActorRefFrom } from "xstate"
-import { ButtonCustom } from "../../../../components/Button"
+import { ButtonCustom } from "../../../../components/Button/ButtonCustom"
 import { EmptyIcon } from "../../../../components/EmptyIcon"
 import { Form } from "../../../../components/Form"
 import { FieldComboInput } from "../../../../components/Form/FieldComboInput"
 import { WithdrawIntentCard } from "../../../../components/IntentCard/WithdrawIntentCard"
 import { NetworkIcon } from "../../../../components/Network/NetworkIcon"
 import { Select } from "../../../../components/Select/Select"
-import { useModalController } from "../../../../hooks"
+import { useModalController } from "../../../../hooks/useModalController"
 import { useTokensStore } from "../../../../providers/TokensStoreProvider"
 import { ModalType } from "../../../../stores/modalStore"
-import { ChainType } from "../../../../types"
 import type {
   BaseTokenInfo,
   SupportedChainName,
   UnifiedTokenInfo,
 } from "../../../../types/base"
+import { ChainType } from "../../../../types/deposit"
 import type { WithdrawWidgetProps } from "../../../../types/withdraw"
-import { isBaseToken } from "../../../../utils"
 import { formatTokenValue } from "../../../../utils/format"
 import { parseUnits } from "../../../../utils/parse"
+import { isBaseToken } from "../../../../utils/token"
 import { validateAddress } from "../../../../utils/validateAddress"
 import type { intentStatusMachine } from "../../../machines/intentStatusMachine"
 import { getPOABridgeInfo } from "../../../machines/poaBridgeInfoActor"
