--- src/features/withdraw/index.ts
@@ -1,1 +1,0 @@
-export { WithdrawWidget } from "./components/WithdrawWidget"
