--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useResize } from "./useResize"
-export { useModalController } from "./useModalController"
