--- src/index.ts
@@ -1,4 +1,0 @@
-export * from "./features"
-export * from "./providers"
-export * from "./types"
-export * from "./utils"
