--- src/providers/index.ts
@@ -1,3 +1,0 @@
-export { SwapWidgetProvider } from "./SwapWidgetProvider"
-export { DepositWidgetProvider } from "./DepositWidgetProvider"
-export { ModalStoreProvider } from "./ModalStoreProvider"
