+++ src/services/depositService.ts
@@ -14,13 +14,13 @@ import {
 } from "viem"
 import { settings } from "../config/settings"
 import { getNearTxSuccessValue } from "../features/machines/getTxMachine"
+import type { SupportedChainName } from "../types/base"
 import {
-  BlockchainEnum,
+  ChainType,
   type SendTransactionEVMParams,
-  type SupportedChainName,
-} from "../types"
-import { ChainType } from "../types"
-import type { Transaction } from "../types/deposit"
+  type Transaction,
+} from "../types/deposit"
+import { BlockchainEnum } from "../types/interfaces"
 import { type DefuseUserId, userAddressToDefuseUserId } from "../utils/defuse"
 import { getEVMChainId } from "../utils/evmChainId"
 import { getDepositAddress, getSupportedTokens } from "./poaBridgeClient"
