+++ src/services/quoteService.ts
@@ -1,5 +1,5 @@
 import { settings } from "../config/settings"
-import type { BaseTokenInfo } from "../types"
+import type { BaseTokenInfo } from "../types/base"
 import { computeTotalBalance } from "../utils/tokenUtils"
 import { quote } from "./solverRelayHttpClient"
 import type { QuoteResponse } from "./solverRelayHttpClient/types"
