--- src/types/index.ts
@@ -1,4 +1,0 @@
-export * from "./base"
-export * from "./swap"
-export * from "./interfaces"
-export * from "./deposit"
